requirements.txt
# Install: pip install -r requirements.txt
requests>=2.31.0
scikit-learn>=1.3.0
sentence-transformers>=2.7.0   # SBERT + E5
hdbscan>=0.8.33
umap-learn>=0.5.6
yfinance>=0.2.40
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
pandas>=2.1.0
jupyter>=1.0.0
ipython>=8.0.0
openai>=1.30.0                 # requires OPENAI_API_KEY